+++ liquid/builtin/tags/case_tag.py
@@ -24,7 +24,7 @@ from liquid.token import Token
 
 if TYPE_CHECKING:
     from liquid import Environment
-    from liquid.context import Context
+    from liquid.context import RenderContext
     from liquid.expression import Expression
 
 # ruff: noqa: D102
@@ -69,7 +69,9 @@ class CaseNode(ast.Node):
 
         return " ".join(buf)
 
-    def render_to_output(self, context: Context, buffer: TextIO) -> Optional[bool]:
+    def render_to_output(
+        self, context: RenderContext, buffer: TextIO
+    ) -> Optional[bool]:
         buf = context.get_buffer(buffer)
         rendered: Optional[bool] = False
 
@@ -87,7 +89,7 @@ class CaseNode(ast.Node):
         return rendered
 
     async def render_to_output_async(
-        self, context: Context, buffer: TextIO
+        self, context: RenderContext, buffer: TextIO
     ) -> Optional[bool]:
         buf = context.get_buffer(buffer)
         rendered: Optional[bool] = False