+++ tests/test_filter.py
@@ -1,9 +1,10 @@
 """Test filter decorators and helpers."""
+
 import asyncio
 from unittest import TestCase
 
-from liquid import Context
 from liquid import Environment
+from liquid import RenderContext
 from liquid.exceptions import FilterArgumentError
 from liquid.filter import int_arg
 from liquid.filter import liquid_filter
@@ -13,13 +14,13 @@ from liquid.filter import with_context
 
 @liquid_filter
 @with_context
-def some_filter(val, arg, *, context: Context) -> str:
+def some_filter(val: str, arg: str, *, context: RenderContext) -> str:
     """Contrived filter function making use of `with_context`."""
-    return val + context.resolve(arg)
+    return val + str(context.resolve(arg))
 
 
 class WithContextTestCase(TestCase):
-    def test_filter_with_context(self):
+    def test_filter_with_context(self) -> None:
         """Test that we can pass the active context to a filter function."""
         env = Environment(strict_filters=True)
         env.add_filter("some", some_filter)
@@ -31,27 +32,27 @@ class WithContextTestCase(TestCase):
 
 
 class IntArgTestCase(TestCase):
-    def test_int_arg_fail(self):
+    def test_int_arg_fail(self) -> None:
         """Test that a suitable exception is raised if we can't cast to an int."""
         with self.assertRaises(FilterArgumentError):
             int_arg("foo")
 
-    def test_int_arg_with_default(self):
+    def test_int_arg_with_default(self) -> None:
         """Test that a default is returned if we can't cast to an int."""
         self.assertEqual(int_arg("foo", 99), 99)
 
 
 class NumArgTestCase(TestCase):
-    def test_num_arg_fail_string(self):
+    def test_num_arg_fail_string(self) -> None:
         """Test that a suitable exception is raised if we can't cast to a number."""
         with self.assertRaises(FilterArgumentError):
             num_arg("foo")
 
-    def test_num_arg_with_default(self):
+    def test_num_arg_with_default(self) -> None:
         """Test that a default is returned if we can't cast to a number."""
         self.assertEqual(num_arg("foo", 99.8), 99.8)
 
-    def test_num_arg_no_default(self):
+    def test_num_arg_no_default(self) -> None:
         """Test that an exception is raised if a default is not given."""
         with self.assertRaises(FilterArgumentError):
             num_arg(object())
@@ -87,7 +88,7 @@ class OtherFilter:
 
 
 class AsyncFilterTestCase(TestCase):
-    def test_call_an_async_filter(self):
+    def test_call_an_async_filter(self) -> None:
         """Test that we can await an async filter."""
         env = Environment(strict_filters=True)
         env.add_filter("greeting", SomeFilter())
@@ -95,7 +96,7 @@ class AsyncFilterTestCase(TestCase):
         result = template.render(you="World")
         self.assertEqual(result, "Hello, World")
 
-    def test_await_an_async_filter(self):
+    def test_await_an_async_filter(self) -> None:
         """Test that we can await an async filter."""
         env = Environment(strict_filters=True)
         env.add_filter("greeting", SomeFilter())
@@ -107,7 +108,7 @@ class AsyncFilterTestCase(TestCase):
         result = asyncio.run(coro())
         self.assertEqual(result, "Goodbye, World")
 
-    def test_await_an_async_filter_with_args(self):
+    def test_await_an_async_filter_with_args(self) -> None:
         """Test that we can await an async filter with arguments."""
         env = Environment(strict_filters=True)
         env.add_filter("greeting", OtherFilter())
@@ -119,7 +120,7 @@ class AsyncFilterTestCase(TestCase):
         result = asyncio.run(coro())
         self.assertEqual(result, "Goodbye, World!")
 
-    def test_await_an_async_filter_with_keyword_args(self):
+    def test_await_an_async_filter_with_keyword_args(self) -> None:
         """Test that we can await an async filter with keyword arguments."""
         env = Environment(strict_filters=True)
         env.add_filter("greeting", OtherFilter())